Two changes to evolve3D
1) Added capability to write out the sequence of intermediate solutions
to the ionization fraction (xh_intermed). This is commented out in this
version, but all the elements are there.
This change required another argument to be added to the call to
evolve3d, which led to a change in C2Ray.F90

2) A floating point number for the average number of subboxes instead
of a rounded integer. Allows for better estimates of CPU time requirements
as it tells you what fraction of the box is being traced.