Changes related to debugging. Comparing the output of rerun of 500Mpc_z50_0_300
Unfortunately, the source model in the original run is flawed in that
it did include LMACH sources, but only if they existed in a cell containing
an HMACH source! However, the comparison did help track down some
mistakes in the new sourceprops.F90 module.

Note, the source model in the new sourceprops.F90 in this version has been
hacked to ignore LMACHs, which is not what we want in the end.

 Changes to be committed:
	modified:   c2ray_parameters.f90
	modified:   nbody_cubep3m.F90
	modified:   sourceprops.F90