Changes related to simplifying the input file. The input_example file now shows what an input file should contain.
I removed the output flags (1st line of old input file). If you want to
specify this, please set the parameter ask_for_streams to .true. and
put the flags on the first line of the input file. By default the code
will now produce the relevant 3D cubes and none of the 1D and 2D ones.

I removed the initial temperature (2nd line of old input file). If you
want to specify	this, please set the parameter ask_for_temperature to
.true. and put the temperature value on the first or second line of the
input file (depending on the value of ask_for_streams). By default the
initial temperature is 1e4.

I removed the name of the clumping file from the input file. This name
is not actually used anymore as we have new clumping algorithms in place.

 On branch factorization
 Changes to be committed:
	modified:   c2ray_parameters.f90
	modified:   inputs/input_example
	modified:   material.F90
	modified:   output.F90