Changes induced by having to run with a redshift list which is not equally spaced in time. The collapsed mass growth source model has now been adjusted to work for this case. The other source models (e.g. Iliev et al.) have NOT been adjusted and will give erroneous results. The role of the time interval for this mass-based source model is problematic in any case.
The changes in the source properties module induced some (comment)
changes in C2Ray.F90, some reordering in the cosmology module as
well as changes in the core makefile. The cosmology module is now
needed by the source properties module and so should be compiled
before. In order to facilitate future changes in the compilation
order of modules, further variables have been introduced in the core
makefile.

 On branch factorization
 Changes to be committed:
	modified:   C2Ray.F90
	modified:   cosmology.F90
	modified:   makefile_core
	modified:   sourceprops.F90